% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_methods.R
\name{addBootSampleObject}
\alias{addBootSampleObject}
\title{Create a bootsrap sample for coefficient estimates}
\usage{
addBootSampleObject(object, ...)
}
\arguments{
\item{object}{an object to add boot object to}

\item{...}{extra parameters to be passed to the \code{boot::boot()} function
other than \code{data} and \code{statistic}.}
}
\value{
\code{object} with additional component \code{"boot"}
}
\description{
Create a boostrap sample from coefficient estimates.
}
\details{
The information in \code{object} is used to prepare the arguments and then
\code{boot} is called to generate the bootstrap sample.
The bootstrap sample is stored in \code{object} as component \code{"boot"}.
Arguments in \code{"..."} can be used customise the \code{boot()} call.
}
\examples{
## see renewal_methods
}
\seealso{
\code{\link{renewal_methods}}
}
