% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renewal_methods.R
\name{addBootSampleObject.renewal}
\alias{addBootSampleObject.renewal}
\title{Create a bootsrap sample for coefficient estimates}
\usage{
addBootSampleObject.renewal(object, R = 100, ...)
}
\arguments{
\item{object}{object returned by one of the count-modeling functions}

\item{R}{numeric, number of bootsrap samples to generate.}

\item{...}{not used.}
}
\value{
\code{object} with additional component \code{"boot"}
}
\description{
Create a boostrap sample from coefficient estimates.
}
\details{
The information in \code{object} is used to prepare the arguments and then
\code{boot} is called to generate the bootstrap sample.
The bootstrap sample is stored in \code{object} as component \code{"boot"}.
}
\examples{
## see renewal_methods
}

