% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anc.R
\name{CountrFormula}
\alias{CountrFormula}
\title{Create a formula for renewalCount}
\usage{
CountrFormula(response, ...)
}
\arguments{
\item{response}{the formula for the "main" parameter. It also specifies the
response variable.}

\item{...}{additional arguments for the ancilliary parameters.}
}
\value{
a Formula object suitable for argument \code{formula} of
    \code{renewalCount()}.
}
\description{
Create a formula for renewalCount
}
