% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dWeibullgammaCount_mat}
\alias{dWeibullgammaCount_mat}
\title{Univariate Weibull Count Probability with gamma heterogeneity}
\usage{
dWeibullgammaCount_mat(x, shape, r, alpha, time = 1, logFlag = FALSE,
  jmax = 100L)
}
\arguments{
\item{r, alpha}{double gamma parameters}
}
\description{
Univariate Weibull Count Probability with gamma heterogeneity
}
\keyword{internal}
