% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dWeibullgammaCount_mat_Covariates}
\alias{dWeibullgammaCount_mat_Covariates}
\title{Univariate Weibull Count Probability with gamma and
covariate heterogeneity}
\usage{
dWeibullgammaCount_mat_Covariates(x, cc, r, alpha, Xcovar, beta, t = 1,
  logFlag = FALSE, jmax = 100L)
}
\arguments{
\item{x, cc, t, logFlag, jmax}{TODO
keywords internal}

\item{r}{numeric shape of the gamma distribution}

\item{alpha}{numeric rate of the gamma distribution}

\item{Xcovar}{matrix covariates value}

\item{beta}{numeric vector of slopes}
}
\description{
Univariate Weibull Count Probability with gamma and
covariate heterogeneity
}
