% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convCount_moments.R
\name{evCount_conv_user}
\alias{evCount_conv_user}
\title{renewal count expected value and variance (user)}
\usage{
evCount_conv_user(xmax, distPars, extrapolPars, survR, method = c("dePril",
  "direct", "naive"), nsteps = 100, time = 1, extrap = TRUE)
}
\arguments{
\item{xmax}{unsigned integer maximum count to be used.}

\item{distPars}{TODO}

\item{extrapolPars}{ma::vec of length 2. The extrapolation values.}

\item{survR}{Rcpp::Function user passed survival function; should have the
signature \code{function(t, distPars)} where \code{t} is a real number (>0)
where the survival function is evaluated and \code{distPars} is a list of
distribution parameters. It should return a double value.}

\item{method}{TODO}

\item{nsteps}{unsiged integer number of steps used to compute the integral.}

\item{time}{double time at wich to compute the probabilities. Set to 1 by
default.}

\item{extrap}{logical if \code{TRUE}, Richardson extrapolation will be
applied to improve accuracy.}
}
\value{
List named vector with ExpectedValue and Variance
}
\description{
Expcted value and varinace of the renewal count process computed numerically
}
\examples{
## see evCount_conv_bi
}

