% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renewal_methods.R
\name{extractAIC.renewal}
\alias{extractAIC.renewal}
\title{extractAIC method for class renewal}
\usage{
\method{extractAIC}{renewal}(fit, scale, k = 2, ...)
}
\arguments{
\item{fit}{an object from class \code{renewal}}

\item{scale}{TODO}

\item{k}{TODO}

\item{...}{not used}
}
\description{
extractAIC method for class renewal
}
\examples{
## see renewal_methods
}

