% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football}
\alias{football}
\title{Football data}
\format{a data.frame with 6 columns and 1104 observations:
\describe{
\item{\code{seasonId}}{integer season identifier (year of the first month of competition).}
\item{\code{gameDate}}{POSIXct game date and time.}
\item{\code{homeTeam,awayTeam}}{character home and away team name}.
\item{\code{homeTeamGoals,awayTeamGoals}}{integer number of goals scored by the home
and the away team.}
}}
\usage{
football
}
\description{
Final scores of all matches in the English Premier League from seasons
2009/2010 to 2016/2017.
}
\details{
The data were collected from \url{http://www.football-data.co.uk/englandm.php}
and slightly formatted and simplified.
}
\keyword{datasets}
