% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_tools.R
\name{frequency_plot}
\alias{frequency_plot}
\title{Plot a frequency chart}
\usage{
frequency_plot(count_labels, actual, pred, colours)
}
\arguments{
\item{count_labels}{character, labels to be used.}

\item{actual}{numeric, the observed probabilities for the different count
specified in \code{count_labels}.}

\item{pred}{data.frame of predicted values. Should have the same number of
rows as actual and one column per model. The columns' names will be used
as labels for the different models.}

\item{colours}{character vector of colour codes with \code{length}
\code{ncol(pred)} + 2.}
}
\description{
Plot a frequency chart to compare actual and predicted values.
}
\details{
In order to compare actual and fitted values, a barchart plot is created.  It
is the user's responsibility to provide the count, observed and fitted
values.
}
