% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewalTools.R
\name{getParNames}
\alias{getParNames}
\title{return the name of the distribution parameters}
\usage{
getParNames(dist, ...)
}
\arguments{
\item{dist}{character name of the distribution}

\item{...}{parameters to pass when dist == "custom"}
}
\value{
character vector with the name of the distribution parameters
}
\description{
return the name of the distribution parameters
}

