% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpredict.R
\name{prob_predict}
\alias{prob_predict}
\title{Probability predictions}
\usage{
prob_predict(obj, ...)
}
\arguments{
\item{obj}{object from a supported class.}

\item{...}{further arguments for methods}
}
\description{
Compute probability predictions for discrete distribution models.
}
\details{
This is an emergency replacement for pscl::predprob, since package 'pscl is
scheduled for removal from CRAN (in a couple of weeks) at the time of writing
this (2024-01-13).

The function is exported but currently marked as internal, since the
arguments may be changed. When it is stable, it will be documented fully.
}
\keyword{internal}
