% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewalTools.R
\name{renewal.convPars}
\alias{renewal.convPars}
\title{Creates the convolution inputs setting}
\usage{
renewal.convPars(convPars, dist)
}
\arguments{
\item{convPars}{a list of convolution parameters arguments with slots
\code{nsteps}, \code{extrap} and \code{convMethod}.
See \code{dCount_conv_bi}. If NULL, default parameters will be applied.}

\item{dist}{TODO}
}
\value{
list convolution inputs.
}
\description{
Check and creates the convolution inputs list
}

