% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_tools.R
\name{renewal.seriesPars}
\alias{renewal.seriesPars}
\title{Creates the series expansion inputs setting}
\usage{
renewal.seriesPars(seriesPars, long = FALSE)
}
\arguments{
\item{seriesPars}{list series expansion input parameters with slots
\code{terms} (number of terms in the series expansion), \code{iter}
(number of iteration in the accelerated series expansion algorithm) and
\code{eps} (tolerance in the accelerated series expansion algorithm),
Only used if \code{dist = "weibull"} and \code{weiMethod =
c("series_mat", "series_acc")}.}

\item{long}{TODO}
}
\value{
list series expansion inputs.
}
\description{
Check and creates the series expansion inputs list.
}
\keyword{internal}
