% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_tools.R
\name{renewal.weiMethod}
\alias{renewal.weiMethod}
\title{Check weibull computation algorithm}
\usage{
renewal.weiMethod(weiMethod)
}
\arguments{
\item{weiMethod}{character, desired  weibull method.}
}
\value{
a valid weibull computation method.
}
\description{
Check weibull computation algorithm.
}
\keyword{internal}
