% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_methods.R
\name{residuals_plot}
\alias{residuals_plot}
\title{Method to visualise the residuals}
\usage{
residuals_plot(object, type, ...)
}
\arguments{
\item{object}{object returned by one of the count-modeling functions}

\item{type}{character type of residuals to be used.}

\item{...}{further arguments for methods.}
}
\description{
A method to visualise the residuals
}
