% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renewal_methods.R
\name{se.coef}
\alias{se.coef}
\title{Extract Standard Errors of Model Coefficients}
\usage{
se.coef(object, parm, type, ...)
}
\arguments{
\item{object}{object returned by one of the count-modeling functions}

\item{parm}{parameters name or index}

\item{type}{type of standard error. User can choose between asymtotic
normal standard errors (\code{asymptotic}) or bootsrap (\code{boot}).}

\item{...}{TODO}
}
\value{
a named numeric vector
}
\description{
These functions extract standard errors of model coefficients
from objects returned by count-modeling functions.
}
\details{
When bootsrap standard error are required, the function cheks for the
bootsrap sample in \code{object}. If it is not found, the bootsrap
sample is created and a warning is sent.
}

