% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renewal_methods.R
\name{se.coef.renewal}
\alias{se.coef.renewal}
\title{Extract Standard Errors of Model Coefficients}
\usage{
\method{se.coef}{renewal}(object, parm, type = c("asymptotic", "boot"), ...)
}
\arguments{
\item{object}{object returned by \code{renewal}.}

\item{parm}{parameters name or index}

\item{type}{type of standard error. User can choose between asymtotic
normal standard errors (\code{asymptotic}) or bootsrap (\code{boot}).}

\item{...}{TODO}
}
\value{
an named numeric vector
}
\description{
These functions extract standard errors of model coefficients
from objects returned by \code{renewal} functions.
}
\details{
When bootsrap standard error are required, the function cheks for the
bootsrap sample in \code{object}. If it is not found, the bootsrap
sample is created and a warning is sent.
}
\examples{
## see examples for renewal_methods
}

