% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertility}
\alias{fertility}
\title{fertility data}
\format{A data frame with 11 variables:
\describe{
\item{\code{Y}}{response variable; number of children per woman (integer).}
\item{\code{GERMAN}}{nationality; dummy variable German/non-German.}
\item{\code{EDU}}{education measured as years of schooling.}
\item{\code{VOC}}{vocational training; dummy}
\item{\code{UNI}}{university education; dummy}
\item{\code{CATH}}{catholic; dummy}
\item{\code{PROT}}{protestant; dummy}
\item{\code{MUSL}}{muslim; dummy}
\item{\code{RURAL}}{rural; dummy}
\item{\code{YEAR_OF_B}}{year of birth; integer}
\item{\code{AGEMARR}}{age at marriage; integer}
}

For further details, see Winlemann(1995).}
\usage{
fertility
}
\description{
Fertility data analysed by Winlemann(1995). The data comes from the scond
(1985) wave of German Socio-Economic Panel. The sample sample is formed by
1,243 women aged 44 or older in 1985. The response variable is the number of
child per woman and a number of covariates described in more details below.
}
\keyword{datasets}

