% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_tools.R
\name{getParNames}
\alias{getParNames}
\title{Return the names of distribution parameters}
\usage{
getParNames(dist, ...)
}
\arguments{
\item{dist}{character, name of the distribution.}

\item{...}{parameters to pass when dist == "custom".}
}
\value{
character vector with the names of the distribution parameters.
}
\description{
Return the names of the parameters of a count distribution.
}
