% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal_cstr.R
\name{renewal}
\alias{renewal}
\title{Deprecated old name of \code{renewalCount()}}
\usage{
renewal(...)
}
\arguments{
\item{...}{same arguments as \code{renewalCount()}.}
}
\description{
Deprecated old name of \code{renewalCount()}
}
\details{
This function is now renamed to \code{renewalCount()}.
}
\keyword{internal}
